requirements.txt
# Base ----------------------------------------
sqlalchemy==2.0.36
pandas==2.2.3
python-dotenv==0.20.0
psycopg2-binary==2.9.10
tabulate==0.9.0

numpy>=1.18.5
opencv-python>=4.1.2
matplotlib>=3.2.2
ipyleaflet>=0.14.0
sympy>=1.10.1
jupyter==1.*

# Machine learning ----------------------------
scikit-learn==1.6.0
